requirements.txt
streamlit>=1.37
openai>=1.30
pandas>=2.2
Pillow>=10.0
openpyxl>=3.1
